Move all components for the MainMenu into a folder together
