+++ src/renderer/components/MainMenu/ChatMenuItem.js
@@ -20,7 +20,7 @@ import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemText from "@material-ui/core/ListItemText";
 import ChatIcon from "@material-ui/icons/Chat";
-import i18n from "../i18n";
+import i18n from "../../i18n";
 
 export default function ChatMenuItem({ onClick }) {
   return (
