+++ src/renderer/components/MainMenu/ExitMenuItem.js
@@ -19,16 +19,16 @@ import React from "react";
 import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemText from "@material-ui/core/ListItemText";
-import HighlightIcon from "@material-ui/icons/Highlight";
-import i18n from "../i18n";
+import ExitToAppIcon from "@material-ui/icons/ExitToApp";
+import i18n from "../../i18n";
 
-export default function ColormapMenuItem({ selected, onClick }) {
+export default function ExitMenuItem({ onClick }) {
   return (
-    <ListItem button selected={selected} onClick={onClick}>
+    <ListItem button onClick={onClick}>
       <ListItemIcon>
-        <HighlightIcon />
+        <ExitToAppIcon />
       </ListItemIcon>
-      <ListItemText primary={i18n.app.menu.colormapEditor} />
+      <ListItemText primary={i18n.app.menu.exit} />
     </ListItem>
   );
 }
