+++ src/renderer/components/MainMenu/FlashMenuItem.js
@@ -19,16 +19,16 @@ import React from "react";
 import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemText from "@material-ui/core/ListItemText";
-import SettingsIcon from "@material-ui/icons/Settings";
-import i18n from "../i18n";
+import CloudUploadIcon from "@material-ui/icons/CloudUpload";
+import i18n from "../../i18n";
 
-export default function SettingsMenuItem({ selected, onClick }) {
+export default function FlashMenuItem({ selected, onClick }) {
   return (
     <ListItem button selected={selected} onClick={onClick}>
       <ListItemIcon>
-        <SettingsIcon />
+        <CloudUploadIcon />
       </ListItemIcon>
-      <ListItemText primary={i18n.app.menu.settings} />
+      <ListItemText primary={i18n.app.menu.firmwareUpdate} />
     </ListItem>
   );
 }
