+++ src/renderer/components/MainMenu/KeyboardSelectMenuItem.js
