+++ src/renderer/components/MainMenu/KeymapMenuItem.js
@@ -20,7 +20,7 @@ import ListItem from "@material-ui/core/ListItem";
 import ListItemIcon from "@material-ui/core/ListItemIcon";
 import ListItemText from "@material-ui/core/ListItemText";
 import KeyboardIcon from "@material-ui/icons/Keyboard";
-import i18n from "../i18n";
+import i18n from "../../i18n";
 
 export default function KeymapMenuItem({ selected, onClick }) {
   return (
